Day 22 | Personal Expense Intelligence | fix: prevent users from accessing or modifying others expenses

+++ week-03-challenge/PersonalExpenseIntelligence/server/controllers/expense.controller.js
@@ -1,9 +1,7 @@
 const {Expense} = require('../models/expense.model')
 
-//Add user authentication
 
 const createExpense = (req, res) => {
-    console.log(req.body)
     const { amount, description, category } = req.body
     Expense.create({
         userId: req.user.id,
@@ -13,35 +11,55 @@ const createExpense = (req, res) => {
     })
         .then(newExpense => {
             res.json(newExpense)
-            console.log('New expense created: ', newExpense)
         })
         .catch(err => res.status(400).json(err))
 }
 
 const getAllExpenses = (req, res) => {
-    Expense.find({})
+    Expense.find({ userId: req.user.id })
         .then(allExpenses => res.json(allExpenses))
         .catch(err => res.status(400).json(err))
 }
 
 const getOneExpense = (req, res) => {
     const { id } = req.params
-    Expense.findOne({ _id: id })
-        .then(oneExpense => res.json(oneExpense))
+    Expense.findOne({ _id: id, userId: req.user.id })
+        .then(oneExpense => {
+            if (!oneExpense) {
+                return res.status(404).json({ message: 'Expense not found' })
+            }
+            res.json(oneExpense)
+        })
         .catch(err => res.status(500).json(err))
 }
 
 const updateExpense = (req, res) => {
     const { id } = req.params
-    Expense.findOneAndUpdate({_id: id}, req.body, {  returnDocument: 'after', runValidators: true })
-        .then(updatedExpense => res.json(updatedExpense))
+    Expense.findOneAndUpdate(
+        {_id: id, userId: req.user.id},
+        req.body,
+        {  returnDocument: 'after', runValidators: true }
+    )
+        .then(updatedExpense => {
+            if (!updatedExpense) {
+                return res.status(404).json({ message: 'Expense not found or not yours' })
+            }            
+            res.json(updatedExpense)
+
+        })
         .catch(err => res.status(400).json(err))
 }
 
 const deleteExpense = (req, res) => {
     const { id } = req.params
-    Expense.deleteOne({_id: id})
-        .then(deletedExpense => res.json(deletedExpense))
+
+    Expense.deleteOne({_id: id, userId: req.user.id })
+        .then(result => {
+            if (result.deletedCount === 0){
+                return res.status(404).json({ message: 'Expense not found or not yours' })
+            }
+            res.json({ message: 'Expense deleted' })
+        })
         .catch(err => res.status(400).json(err))
 }
